provider: shim layer for older clients wrt StorageClaim resources
